fix(setup):  rewrote the initial setup
In the old version in the initial setup the program has to be started twice. Once for creating the default config file and the second time for editing it.

In the new version this is combined. If the programm is started the first time it creates the default config and asks the user about the settings before saving.